extract testkit logic from to separate files
